[DOWNLOAD] Rewrite Download
...and fix a typo in download.tiles

...i might divide the download.panorama function into two, or just remove it all idk lol
